+++ src/modules/userstyles.ts
@@ -549,7 +549,7 @@ function parseUserCSS(code: string): ParseResult {
       }
     } else if (key === 'match' && value) {
       matchPatterns.push(value);
-    } else if (key in meta) {
+    } else if (Object.prototype.hasOwnProperty.call(meta, key)) {
       meta[key] = value;
     }
   }